+++ PgRoutiner/Settings/BuildFormatedSettings.cs
@@ -12,15 +12,15 @@ namespace PgRoutiner
         {
             StringBuilder sb = new();
 
-            void AddSectionComment(string header, string helpUrl, string tip = null)
+            void AddSectionComment(string header, string helpUrl, params string[] tips)
             {
                 sb.AppendLine($"    /*");
                 sb.AppendLine($"      {header}");
-                if (tip != null)
+                foreach(var tip in tips)
                 {
                     sb.AppendLine($"      {tip}");
                 }
-                sb.AppendLine($"      For more info, see: {helpUrl}");
+                sb.AppendLine($"      - For more info see: {helpUrl}");
                 sb.AppendLine($"    */");
             }
 
@@ -92,8 +92,12 @@ namespace PgRoutiner
             }
 
             AddSectionComment(
-                "General settings", 
-                "https://github.com/vb-consulting/PgRoutiner/wiki/1.-WORKING-WITH-SETTINGS#general-settings");
+                "General settings:", 
+                "https://github.com/vb-consulting/PgRoutiner/wiki/1.-WORKING-WITH-SETTINGS#general-settings",
+                $"- Use \"{ConnectionArgs.Alias}\" or \"--{ConnectionArgs.Original.ToKebabCase()}\" option to set working connection from the command line.",
+                $"- Use \"{SchemaArgs.Alias}\" or \"--{SchemaArgs.Original.ToKebabCase()}\" option to set schema similar to expression from the command line.",
+                $"- Use \"{ExecuteArgs.Alias}\" or \"--{ExecuteArgs.Original.ToKebabCase()}\" option to execute SQL file or PSQL command on your current connection  from the command line.",
+                $"- Use \"{DumpArgs.Alias}\" or \"--{DumpArgs.Original.ToKebabCase()}\" switch to redirect all outputs to the command line.");
             AddEntry(nameof(Connection), Value.Connection);
             AddEntry(nameof(Schema), Value.Schema);
             AddEntry(nameof(Execute), Value.Execute);
@@ -109,7 +113,10 @@ namespace PgRoutiner
             AddSectionComment(
                 "Routines data-access extensions code-generation", 
                 "https://github.com/vb-consulting/PgRoutiner/wiki/2.-WORKING-WITH-ROUTINES#routines-data-access-extensions-code-generation-settings",
-                $"Use \"{RoutinesArgs.Alias}\" or \"{RoutinesArgs.Name}\" switch to run routines data-access extensions code-generation from the command line.");
+                $"- Use \"{RoutinesArgs.Alias}\" or \"--{RoutinesArgs.Original.ToKebabCase()}\" switch to run routines data-access extensions code-generation from the command line.",
+                $"- Use \"{OutputDirArgs.Alias}\" or \"--{OutputDirArgs.Original.ToKebabCase()}\" option to set the output dir for the generated code from the command line.",
+                $"- Use \"{RoutinesOverwriteArgs.Alias}\" or \"--{RoutinesOverwriteArgs.Original.ToKebabCase()}\" switch to set the overwerite mode for the generated code from the command line.",
+                $"- Use \"{ModelDirArgs.Alias}\" or \"--{ModelDirArgs.Original.ToKebabCase()}\" option to set the custom models output dir for the generated code from the command line.");
             AddEntry(nameof(Routines), Value.Routines);
             AddEntry(nameof(OutputDir), Value.OutputDir);
             AddEntry(nameof(RoutinesOverwrite), Value.RoutinesOverwrite);
@@ -130,7 +137,8 @@ namespace PgRoutiner
             AddSectionComment(
                 "Unit tests code-generation settings", 
                 "https://github.com/vb-consulting/PgRoutiner/wiki/3.-WORKING-WITH-UNIT-TESTS#unit-tests-code-generation-settings",
-                $"Use \"{UnitTestsArgs.Alias}\" or \"--{UnitTestsArgs.Original.ToKebabCase()}\" switch to run unit tests code-generation from the command line.");
+                $"- Use \"{UnitTestsArgs.Alias}\" or \"--{UnitTestsArgs.Original.ToKebabCase()}\" switch to run unit tests code-generation from the command line.",
+                $"- Use \"{UnitTestsDirArgs.Alias}\" or \"--{UnitTestsDirArgs.Original.ToKebabCase()}\" option to set the output dir for the generated unit test project from the command line.");
             AddEntry(nameof(UnitTests), Value.UnitTests);
             AddEntry(nameof(UnitTestsDir), Value.UnitTestsDir);
             AddEntry(nameof(UnitTestsAskRecreate), Value.UnitTestsAskRecreate);
@@ -139,7 +147,10 @@ namespace PgRoutiner
             AddSectionComment(
                 "Schema dump script settings", 
                 "https://github.com/vb-consulting/PgRoutiner/wiki/4.-WORKING-WITH-SCHEMA-DUMP-SCRIPT#schema-dump-script-settings",
-                $"Use \"{SchemaDumpArgs.Alias}\" or \"--{SchemaDumpArgs.Original.ToKebabCase()}\" switch to run schema script dump from the command line.");
+                $"- Use \"{SchemaDumpArgs.Alias}\" or \"--{SchemaDumpArgs.Original.ToKebabCase()}\" switch to run schema script dump from the command line.",
+                $"- Use \"{SchemaDumpFileArgs.Alias}\" or \"--{SchemaDumpFileArgs.Original.ToKebabCase()}\" option to set generated schema file name from the command line.",
+                $"- Use \"{SchemaDumpOverwriteArgs.Alias}\" or \"--{SchemaDumpOverwriteArgs.Original.ToKebabCase()}\" switch to set the overwrite mode for the generated schema file from the command line.",
+                $"- Use \"--{nameof(SchemaDumpPrivileges).ToKebabCase()}\" switch to include object privileges in schema file from the command line.");
             AddEntry(nameof(SchemaDump), Value.SchemaDump);
             AddEntry(nameof(SchemaDumpFile), Value.SchemaDumpFile);
             AddEntry(nameof(SchemaDumpOverwrite), Value.SchemaDumpOverwrite);
@@ -154,7 +165,9 @@ namespace PgRoutiner
             AddSectionComment(
                 "Data dump script settings", 
                 "https://github.com/vb-consulting/PgRoutiner/wiki/5.-WORKING-WITH-DATA-DUMP-SCRIPT#data-dump-script-settings",
-                $"Use \"{DataDumpArgs.Alias}\" or \"--{DataDumpArgs.Original.ToKebabCase()}\" switch to run data script dump from the command line.");
+                $"- Use \"{DataDumpArgs.Alias}\" or \"--{DataDumpArgs.Original.ToKebabCase()}\" switch to run data script dump from the command line.",
+                $"- Use \"{DataDumpFileArgs.Alias}\" or \"--{DataDumpFileArgs.Original.ToKebabCase()}\" option to set generated data script file name from the command line.",
+                $"- Use \"{DataDumpOverwriteArgs.Alias}\" or \"--{DataDumpOverwriteArgs.Original.ToKebabCase()}\" switch to set the overwrite mode for the generated data script file from the command line.");
             AddEntry(nameof(DataDump), Value.DataDump);
             AddEntry(nameof(DataDumpFile), Value.DataDumpFile);
             AddEntry(nameof(DataDumpOverwrite), Value.DataDumpOverwrite);
@@ -165,9 +178,11 @@ namespace PgRoutiner
 
             sb.AppendLine();
             AddSectionComment(
-                "Object file tree settings", 
+                "Object file tree settings",
                 "https://github.com/vb-consulting/PgRoutiner/wiki/6.-WORKING-WITH-OBJECT-FILES-TREE#object-file-tree-settings",
-                $"Use \"{DbObjectsArgs.Alias}\" or \"--{DbObjectsArgs.Original.ToKebabCase()}\" switch to run object files tree dump from the command line.");
+                $"- Use \"{DbObjectsArgs.Alias}\" or \"--{DbObjectsArgs.Original.ToKebabCase()}\" switch to run object files tree dump from the command line.",
+                $"- Use \"{DbObjectsDirArgs.Alias}\" or \"--{DbObjectsDirArgs.Original.ToKebabCase()}\" option to set the root output dir from the command line.",
+                $"- Use \"{DbObjectsOverwriteArgs.Alias}\" or \"--{DbObjectsOverwriteArgs.Original.ToKebabCase()}\" switch to set the overwrite mode for the generated files from the command line.");
             AddEntry(nameof(DbObjects), Value.DbObjects);
             AddEntry(nameof(DbObjectsDir), Value.DbObjectsDir);
             AddEntry(nameof(DbObjectsOverwrite), Value.DbObjectsOverwrite);
@@ -184,7 +199,9 @@ namespace PgRoutiner
             AddSectionComment(
                 "Markdown (MD) database dictionaries settings", 
                 "https://github.com/vb-consulting/PgRoutiner/wiki/7.-WORKING-WITH-MARKDOWN-DATABASE-DICTIONARIES#markdown-md-database-dictionaries-settings",
-                $"Use \"{MarkdownArgs.Alias}\" or \"--{MarkdownArgs.Original.ToKebabCase()}\" switch to run markdown (MD) database dictionary file from the command line.");
+                $"- Use \"{MarkdownArgs.Alias}\" or \"--{MarkdownArgs.Original.ToKebabCase()}\" switch to run markdown (MD) database dictionary file from the command line.",
+                $"- Use \"{MdFileArgs.Alias}\" or \"--{MdFileArgs.Original.ToKebabCase()}\" option to set generated dictionary file name from the command line.",
+                $"- Use \"{CommitMdArgs.Alias}\" or \"--{CommitMdArgs.Original.ToKebabCase()}\" switch to run commit changes in comments from the MD fiole back to the database from the command line.");
             AddEntry(nameof(Markdown), Value.Markdown);
             AddEntry(nameof(MdFile), Value.MdFile);
             AddEntry(nameof(MdOverwrite), Value.MdOverwrite);
@@ -199,7 +216,7 @@ namespace PgRoutiner
             AddSectionComment(
                 "PSQL command-line utility settings", 
                 "https://github.com/vb-consulting/PgRoutiner/wiki/8.-WORKING-WITH-PSQL#psql-command-line-utility-settings",
-                $"Use \"{PsqlArgs.Alias}\" or \"--{PsqlArgs.Original.ToKebabCase()}\" switch to open PSQL command-line utility from the command line.");
+                $"- Use \"{PsqlArgs.Alias}\" or \"--{PsqlArgs.Original.ToKebabCase()}\" switch to open PSQL command-line utility from the command line.");
             AddEntry(nameof(Psql), Value.Psql);
             AddEntry(nameof(PsqlTerminal), Value.PsqlTerminal);
             AddEntry(nameof(PsqlCommand), Value.PsqlCommand);
@@ -209,7 +226,8 @@ namespace PgRoutiner
             AddSectionComment(
                 "Diff scripts settings", 
                 "https://github.com/vb-consulting/PgRoutiner/wiki/9.-WORKING-WITH-DIFF-SCRIPTS#diff-scripts-settings",
-                $"Use \"{DiffArgs.Alias}\" or \"--{DiffArgs.Original.ToKebabCase()}\" diff script generation from the command line.");
+                $"- Use \"{DiffArgs.Alias}\" or \"--{DiffArgs.Original.ToKebabCase()}\" diff script generation from the command line.",
+                $"- Use \"{DiffArgs.Alias}\" or \"--{DiffArgs.Original.ToKebabCase()}\" option to set target connection for the diff script generator from the command line.");
             AddEntry(nameof(Diff), Value.Diff);
             AddEntry(nameof(DiffTarget), Value.DiffTarget);
             AddEntry(nameof(DiffFilePattern), Value.DiffFilePattern);